+++ Craftsman/Builders/Features/CommandUpdateRecordBuilder.cs
@@ -92,7 +92,7 @@ public static class {className}
                 .FirstOrDefaultAsync({entity.Lambda} => {entity.Lambda}.{primaryKeyPropName} == request.{primaryKeyPropName}, cancellationToken);
 
             if ({updatedEntityProp} == null)
-                throw new KeyNotFoundException();
+                throw new NotFoundException(""{entity.Name}"", request.{primaryKeyPropName});
 
             _mapper.Map(request.{commandProp}, {updatedEntityProp});