+++ Craftsman/Builders/Features/QueryGetListBuilder.cs
@@ -83,9 +83,6 @@ public static class {className}
 
         public async Task<PagedList<{readDto}>> Handle({queryListName} request, CancellationToken cancellationToken)
         {{
-            if (request.QueryParameters == null)
-                throw new ApiException(""Invalid query parameters."");
-
             var collection = _db.{entity.Plural}
                 as IQueryable<{entity.Name}>;