+++ Craftsman/Builders/WebApiAppExtensionsBuilder.cs
@@ -8,41 +8,6 @@
 
     public class WebApiAppExtensionsBuilder
     {
-        public static void CreateErrorHandlerWebApiAppExtension(string solutionDirectory, string projectBaseName, IFileSystem fileSystem)
-        {
-            var classPath = ClassPathHelper.WebApiApplicationExtensionsClassPath(solutionDirectory, $"ErrorHandlerAppExtension.cs", projectBaseName);
-
-            if (!fileSystem.Directory.Exists(classPath.ClassDirectory))
-                fileSystem.Directory.CreateDirectory(classPath.ClassDirectory);
-
-            if (fileSystem.File.Exists(classPath.FullClassPath))
-                throw new FileAlreadyExistsException(classPath.FullClassPath);
-
-            using (var fs = fileSystem.File.Create(classPath.FullClassPath))
-            {
-                var data = "";
-                data = GetErrorHandlerAppExtensionText(classPath.ClassNamespace, solutionDirectory, projectBaseName);
-                fs.Write(Encoding.UTF8.GetBytes(data));
-            }
-        }
-
-        public static string GetErrorHandlerAppExtensionText(string classNamespace, string solutionDirectory, string projectName)
-        {
-            var webApiClassPath = ClassPathHelper.WebApiMiddlewareClassPath(solutionDirectory, "", projectName);
-            return @$"namespace {classNamespace};
-
-using Microsoft.AspNetCore.Builder;
-using {webApiClassPath.ClassNamespace};
-
-public static class ErrorHandlerAppExtension
-{{
-    public static void UseErrorHandlingMiddleware(this IApplicationBuilder app)
-    {{
-        app.UseMiddleware<ErrorHandlerMiddleware>();
-    }}
-}}";
-        }
-
         public static void CreateSwaggerWebApiAppExtension(string solutionDirectory, SwaggerConfig swaggerConfig, bool addJwtAuthentication, string projectBaseName, IFileSystem fileSystem)
         {
             var classPath = ClassPathHelper.WebApiApplicationExtensionsClassPath(solutionDirectory, $"SwaggerAppExtension.cs", projectBaseName);